cleanup 2: move stuff between files
